Adaptive LASSO Estimator

adaptiveLassoEst() applied the adaptive LASSO to the entries of the sample covariance matrix. The thresholding function is inspired by the penalized regression introduced by Zou (2006) . The thresholding function assigns a weight to each entry of the sample covariance matrix based on its initial value. This weight then determines the relative size of the penalty resulting in larger values being penalized less and reducing bias (Rothman et al. 2009) .

adaptiveLassoEst(dat, lambda, n)

Arguments

dat
A numeric data.frame, matrix, or similar object.

lambda
A non-negative numeric defining the amount of thresholding applied to each element of dat's sample covariance matrix.

n
A non-negative numeric defining the exponent of the adaptive weight applied to each element of dat's sample covariance matrix.

Value

A matrix corresponding to the estimate of the covariance matrix.

References

Rothman AJ, Levina E, Zhu J (2009). “Generalized Thresholding of Large Covariance Matrices.” Journal of the American Statistical Association, 104(485), 177--186. doi: 10.1198/jasa.2009.0101 , https://doi.org/10.1198/jasa.2009.0101.

Zou H (2006). “The Adaptive Lasso and Its Oracle Properties.” Journal of the American Statistical Association, 101(476), 1418--1429. doi: 10.1198/016214506000000735 , https://doi.org/10.1198/016214506000000735.

Examples

adaptiveLassoEst(dat = mtcars, lambda = 0.9, n = 0.9)

#>              mpg         cyl        disp          hp       drat          wt
#> mpg    36.291826  -9.0609940  -633.09474 -320.727511   1.791643  -4.9283324
#> cyl    -9.060994   2.9013065   199.65332  101.918699   0.000000   0.7496917
#> disp -633.094743 199.6533192 15360.79969 6721.158375 -47.038454 107.6720667
#> hp   -320.727511 101.9186995  6721.15838 4700.866530 -16.385270  44.1656064
#> drat    1.791643   0.0000000   -47.03845  -16.385270   0.000000   0.0000000
#> wt     -4.928332   0.7496917   107.67207   44.165606   0.000000   0.1060738
#> qsec    4.298104  -1.4245844   -96.03823  -86.755340   0.000000   0.0000000
#> vs      1.581827   0.0000000   -44.35067  -24.942707   0.000000   0.0000000
#> am      1.322579   0.0000000   -36.53193   -8.198967   0.000000   0.0000000
#> gear    1.722184   0.0000000   -50.77876   -6.203983   0.000000   0.0000000
#> carb   -5.182560   0.9586480    79.05272   83.020954   0.000000   0.0000000
#>            qsec         vs         am       gear      carb
#> mpg    4.298104   1.581827   1.322579   1.722184 -5.182560
#> cyl   -1.424584   0.000000   0.000000   0.000000  0.958648
#> disp -96.038229 -44.350668 -36.531926 -50.778756 79.052723
#> hp   -86.755340 -24.942707  -8.198967  -6.203983 83.020954
#> drat   0.000000   0.000000   0.000000   0.000000  0.000000
#> wt     0.000000   0.000000   0.000000   0.000000  0.000000
#> qsec   2.905253   0.000000   0.000000   0.000000 -1.433437
#> vs     0.000000   0.000000   0.000000   0.000000  0.000000
#> am     0.000000   0.000000   0.000000   0.000000  0.000000
#> gear   0.000000   0.000000   0.000000   0.000000  0.000000
#> carb  -1.433437   0.000000   0.000000   0.000000  2.263526