Cross-Validated Covariance Matrix Estimator Selector

cvCovEst() identifies the optimal covariance matrix estimator from among a set of candidate estimators.

cvCovEst(
  dat,
  estimators = c(linearShrinkEst, thresholdingEst, sampleCovEst),
  estimator_params = list(linearShrinkEst = list(alpha = 0), thresholdingEst =
    list(gamma = 0)),
  cv_loss = cvMatrixFrobeniusLoss,
  cv_scheme = "v_fold",
  mc_split = 0.5,
  v_folds = 10L,
  center = TRUE,
  scale = FALSE,
  parallel = FALSE
)

Arguments

dat

A numeric data.frame, matrix, or similar object.

estimators

A list of estimator functions to be considered in the cross-validated estimator selection procedure.

estimator_params

A named list of arguments corresponding to the hyperparameters of covariance matrix estimators in estimators. The name of each list element should match the name of an estimator passed to estimators. Each element of the estimator_params is itself a named list, with the names corresponding to a given estimator's hyperparameter(s). The hyperparameter(s) may be in the form of a single numeric or a numeric vector. If no hyperparameter is needed for a given estimator, then the estimator need not be listed.

cv_loss

A function indicating the loss function to be used. This defaults to the Frobenius loss, cvMatrixFrobeniusLoss(). An observation-based version, cvFrobeniusLoss(), is also made available. Additionally, the cvScaledMatrixFrobeniusLoss() is included for situations in which dat's variables are of different scales.

cv_scheme

A character indicating the cross-validation scheme to be employed. There are two options: (1) V-fold cross-validation, via "v_folds"; and (2) Monte Carlo cross-validation, via "mc". Defaults to Monte Carlo cross-validation.

mc_split

A numeric between 0 and 1 indicating the proportion of observations to be included in the validation set of each Monte Carlo cross-validation fold.

v_folds

An integer larger than or equal to 1 indicating the number of folds to use for cross-validation. The default is 10, regardless of the choice of cross-validation scheme.

center

A logical indicating whether to center the columns of dat to have mean zero.

scale

A logical indicating whether to scale the columns of dat to have unit variance.

parallel

A logical option indicating whether to run the main cross-validation loop with future_lapply(). This is passed directly to cross_validate().

Value

A list of results containing the following elements:

estimate - A matrix corresponding to the estimate of the optimal covariance matrix estimator.

estimator - A character indicating the optimal estimator and corresponding hyperparameters, if any.

risk_df - A tibble providing the cross-validated risk estimates of each estimator.

cv_df - A tibble providing each estimators' loss over the folds of the cross-validated procedure.

args - A named list containing arguments passed to cvCovEst.

Examples

cvCovEst(
  dat = mtcars,
  estimators = c(
    linearShrinkLWEst, thresholdingEst, sampleCovEst
  ),
  estimator_params = list(
    thresholdingEst = list(gamma = seq(0.1, 0.3, 0.1))
  ),
  center = TRUE,
  scale = TRUE
)
#> $estimate
#>             mpg        cyl       disp         hp       drat         wt
#> mpg   1.0000000 -0.8521620 -0.8475514 -0.7761684  0.6811719 -0.8676594
#> cyl  -0.8521620  1.0000000  0.9020329  0.8324475 -0.6999381  0.7824958
#> disp -0.8475514  0.9020329  1.0000000  0.7909486 -0.7102139  0.8879799
#> hp   -0.7761684  0.8324475  0.7909486  1.0000000 -0.4487591  0.6587479
#> drat  0.6811719 -0.6999381 -0.7102139 -0.4487591  1.0000000 -0.7124406
#> wt   -0.8676594  0.7824958  0.8879799  0.6587479 -0.7124406  1.0000000
#> qsec  0.4186840 -0.5912421 -0.4336979 -0.7082234  0.0000000  0.0000000
#> vs    0.6640389 -0.8108118 -0.7104159 -0.7230967  0.4402785 -0.5549157
#> am    0.5998324 -0.5226070 -0.5912270  0.0000000  0.7127111 -0.6924953
#> gear  0.4802848 -0.4926866 -0.5555692  0.0000000  0.6996101 -0.5832870
#> carb -0.5509251  0.5269883  0.3949769  0.7498125  0.0000000  0.4276059
#>            qsec         vs         am       gear       carb
#> mpg   0.4186840  0.6640389  0.5998324  0.4802848 -0.5509251
#> cyl  -0.5912421 -0.8108118 -0.5226070 -0.4926866  0.5269883
#> disp -0.4336979 -0.7104159 -0.5912270 -0.5555692  0.3949769
#> hp   -0.7082234 -0.7230967  0.0000000  0.0000000  0.7498125
#> drat  0.0000000  0.4402785  0.7127111  0.6996101  0.0000000
#> wt    0.0000000 -0.5549157 -0.6924953 -0.5832870  0.4276059
#> qsec  1.0000000  0.7445354  0.0000000  0.0000000 -0.6562492
#> vs    0.7445354  1.0000000  0.0000000  0.0000000 -0.5696071
#> am    0.0000000  0.0000000  1.0000000  0.7940588  0.0000000
#> gear  0.0000000  0.0000000  0.7940588  1.0000000  0.0000000
#> carb -0.6562492 -0.5696071  0.0000000  0.0000000  1.0000000
#> 
#> $estimator
#> [1] "thresholdingEst, gamma = 0.3"
#> 
#> $risk_df
#> # A tibble: 5 × 3
#>   estimator         hyperparameters      cv_risk
#>   <chr>             <chr>                  <dbl>
#> 1 thresholdingEst   gamma = 0.3             52.4
#> 2 sampleCovEst      hyperparameters = NA    52.6
#> 3 linearShrinkLWEst hyperparameters = NA    52.6
#> 4 thresholdingEst   gamma = 0.1             52.9
#> 5 thresholdingEst   gamma = 0.2             53.4
#> 
#> $cv_df
#> # A tibble: 50 × 4
#>    estimator         hyperparameters       loss  fold
#>    <chr>             <chr>                <dbl> <int>
#>  1 linearShrinkLWEst hyperparameters = NA  44.5     1
#>  2 thresholdingEst   gamma = 0.1           47.6     1
#>  3 thresholdingEst   gamma = 0.2           47.8     1
#>  4 thresholdingEst   gamma = 0.3           46.1     1
#>  5 sampleCovEst      hyperparameters = NA  47.1     1
#>  6 linearShrinkLWEst hyperparameters = NA  37.1     2
#>  7 thresholdingEst   gamma = 0.1           42.5     2
#>  8 thresholdingEst   gamma = 0.2           42.9     2
#>  9 thresholdingEst   gamma = 0.3           41.6     2
#> 10 sampleCovEst      hyperparameters = NA  42.5     2
#> # … with 40 more rows
#> 
#> $args
#> $args$cv_loss
#> <quosure>
#> expr: ^cvMatrixFrobeniusLoss
#> env:  0x7ffcd8d80070
#> 
#> $args$cv_scheme
#> [1] "v_fold"
#> 
#> $args$mc_split
#> [1] 0.5
#> 
#> $args$v_folds
#> [1] 10
#> 
#> $args$center
#> [1] TRUE
#> 
#> $args$scale
#> [1] TRUE
#> 
#> $args$parallel
#> [1] FALSE
#> 
#> 
#> attr(,"class")
#> [1] "cvCovEst"