Multiple Heat Map Plot

cvMultiMelt() visualizes the structure of one or more covariance matrix estimators through a grid of heat maps, where each heat map corresponds to a different estimator.

cvMultiMelt(
  dat,
  estimator,
  stat = "min",
  dat_orig,
  plot_type = "heatmap",
  cv_details,
  has_hypers,
  abs_v = TRUE
)

Arguments

dat
A named list. Specifically, this is the standard output of cvCovEst.

estimator
A character vector specifying one or more classes of estimators to visualize and compare.

stat
A character vector containing the names of various cross-validated risk summary statistics. Estimators corresponding to each statistics will be visualized with a different heatmap. Default is 'min' for minimum cross-validated risk.

dat_orig
The numeric data.frame, matrix, or similar object originally passed to cvCovEst().

plot_type
A character detailing the type of plot. Passed to theme_cvCovEst, defaults to "risk"

cv_details
A character vector summarizing key arguments passed to cvCovEst().

has_hypers
A character vector containing the names of current estimators with hyperparameters.

abs_v
A logical determining if the absolute value of the matrix entries should be displayed versus the signed value. Default is TRUE.

Value

A grid of heat map plots comparing the desired covariance matrix estimators.