Summary Statistics of Cross-Validated Risk by Estimator Class

cvRiskByClass() calculates the following summary statistics for the cross-validated risk within each class of estimator passed to cvCovEst(): minimum, Q1, median, mean, Q3, and maximum. The results are output as a tibble.

cvRiskByClass(dat)

Arguments

dat

The tibble of cross-validated risk calculations which is output by cvCovEst().

Value

tibble with rows corresponding to estimator classes and columns corresponding to each summary statistic.