POET Estimator

poetEst() implements the Principal Orthogonal complEment Thresholding (POET) estimator, a nonparametric, unobserved-factor-based estimator of the covariance matrix (Fan et al. 2013) . The estimator is defined as the sum of the sample covariance matrix' rank-k approximation and its post-thresholding principal orthogonal complement. The hard thresholding function is used here, though others could be used instead.

poetEst(dat, k, lambda)

Arguments

dat

A numeric data.frame, matrix, or similar object.

k

An integer indicating the number of unobserved latent factors. Empirical evidence suggests that the POET estimator is robust to overestimation of this hyperparameter (Fan et al. 2013) . In practice, it is therefore preferable to use larger values.

lambda

A non-negative numeric defining the amount of thresholding applied to each element of sample covariance matrix's orthogonal complement.

Value

A matrix corresponding to the estimate of the covariance matrix.

References

Fan J, Liao Y, Mincheva M (2013). “Large covariance estimation by thresholding principal orthogonal complements.” Journal of the Royal Statistical Society. Series B (Statistical Methodology), 75(4), 603--680. ISSN 13697412, 14679868, https://www.jstor.org/stable/24772450.

Examples

poetEst(dat = mtcars, k = 2L, lambda = 0.1)
#>              mpg         cyl        disp          hp         drat          wt
#> mpg    36.324103  -9.1723790  -633.09721 -320.732056   2.19506351  -5.1166847
#> cyl    -9.172379   3.1895161   199.68181  101.957810  -0.57764791   1.3759564
#> disp -633.097208 199.6818126 15360.79983 6721.151681 -47.07821023 107.7090659
#> hp   -320.732056 101.9578102  6721.15168 4700.866935 -16.45539748  44.2081565
#> drat    2.195064  -0.5776479   -47.07821  -16.455397   0.28588135  -0.3369854
#> wt     -5.116685   1.3759564   107.70907   44.208157  -0.33698538   0.9573790
#> qsec    4.509149  -1.8868548   -96.02405  -86.804075   0.08714073  -0.3054816
#> vs      1.968872  -0.7298387   -44.37791  -24.999372   0.12286905  -0.3019144
#> am      1.803931  -0.4118850   -36.58262   -8.324576   0.13020674  -0.2692619
#> gear    2.135685  -0.6491935   -50.82168   -6.358748   0.19314430  -0.3827046
#> carb   -5.363105   1.4299448    79.09023   83.088899  -0.12824916   0.6757903
#>              qsec           vs           am         gear        carb
#> mpg    4.50914919   1.96887209   1.80393145   2.13568548 -5.36310484
#> cyl   -1.88685484  -0.72983871  -0.41188504  -0.64919355  1.42994485
#> disp -96.02404610 -44.37791164 -36.58261788 -50.82168281 79.09023204
#> hp   -86.80407537 -24.99937226  -8.32457551  -6.35874793 83.08889861
#> drat   0.08714073   0.12286905   0.13020674   0.19314430 -0.12824916
#> wt    -0.30548161  -0.30191440  -0.26926186  -0.38270459  0.67579032
#> qsec   3.19316613   0.67056452  -0.20495968  -0.28040323 -1.89411290
#> vs     0.67056452   0.25403226   0.08132497   0.09646868 -0.38225399
#> am    -0.20495968   0.08132497   0.24899194   0.29233871  0.02326462
#> gear  -0.28040323   0.09646868   0.29233871   0.54435484  0.32661290
#> carb  -1.89411290  -0.38225399   0.02326462   0.32661290  2.60887097