Generic Summary Method for cvCovEst

summary() provides summary statistics regarding the performance of cvCovEst() and can be used for diagnostic plotting.

# S3 method for cvCovEst
summary(
  object,
  dat_orig,
  summ_fun = c("cvRiskByClass", "bestInClass", "worstInClass", "hyperRisk"),
  ...
)

Arguments

object

A named list of class "cvCovEst".

dat_orig

The numeric data.frame, matrix, or similar object originally passed to cvCovEst().

summ_fun

A character vector specifying which summaries to output. See Details for function descriptions.

...

Additional arguments passed to summary()These are not explicitly used and should be ignored by the user.

Value

A named list where each element corresponds to the output of of the requested summaries.

Details

summary() accepts four different choices for the summ_fun argument. The choices are:

"cvRiskByClass" - Returns the minimum, first quartile, median, third quartile, and maximum of the cross-validated risk associated with each class of estimator passed to cvCovEst().

"bestInClass" - Returns the specific hyperparameters, if applicable, of the best performing estimator within each class along with other metrics.

"worstInClass" - Returns the specific hyperparameters, if applicable, of the worst performing estimator within each class along with other metrics.

"hyperRisk" - For estimators that take hyperparameters as arguments, this function returns the hyperparameters associated with the minimum, first quartile, median, third quartile, and maximum of the cross-validated risk within each class of estimator. Each class has its own tibble, which are returned as a list.

Examples

cv_dat <- cvCovEst(
  dat = mtcars,
  estimators = c(
    linearShrinkEst, thresholdingEst, sampleCovEst
  ),
  estimator_params = list(
    linearShrinkEst = list(alpha = seq(0.1, 0.9, 0.1)),
    thresholdingEst = list(gamma = seq(0.1, 0.9, 0.1))
  ),
  center = TRUE,
  scale = TRUE
)

summary(cv_dat, mtcars)
#> $cvRiskByClass
#> # A tibble: 3 × 7
#>   Estimator         Min    Q1 Median    Q3   Max  Mean
#>   <chr>           <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 sampleCovEst     74.3  74.3   74.3  74.3  74.3  74.3
#> 2 linearShrinkEst  73.3  73.4   75.6  85.1  95.5  80.7
#> 3 thresholdingEst  74.4  74.6   75.8  97.8 108.   86.4
#> 
#> $bestInClass
#> # A tibble: 3 × 6
#>   estimator       hyperparameter       cv_risk cond_num sign  sparsity
#>   <chr>           <chr>                  <dbl>    <dbl> <chr>    <dbl>
#> 1 linearShrinkEst alpha = 0.8             73.3    64400 PD        0   
#> 2 sampleCovEst    hyperparameters = NA    74.3   473000 PD        0   
#> 3 thresholdingEst gamma = 0.3             74.4   -32600 IND       0.22
#> 
#> $worstInClass
#> # A tibble: 3 × 6
#>   estimator       hyperparameter       cv_risk cond_num sign  sparsity
#>   <chr>           <chr>                  <dbl>    <dbl> <chr>    <dbl>
#> 1 sampleCovEst    hyperparameters = NA    74.3   473000 PD        0   
#> 2 linearShrinkEst alpha = 0.1             95.5     2060 PD        0   
#> 3 thresholdingEst gamma = 0.9            108.    -11400 IND       0.43
#> 
#> $hyperRisk
#> $hyperRisk$thresholdingEst
#> # A tibble: 5 × 3
#>   hyperparameters cv_risk stat  
#>   <chr>           <chr>   <chr> 
#> 1 gamma = 0.3     74      min   
#> 2 gamma = 0.1     75      Q1    
#> 3 gamma = 0.4     76      median
#> 4 gamma = 0.7     98      Q3    
#> 5 gamma = 0.9     108     max   
#> 
#> $hyperRisk$linearShrinkEst
#> # A tibble: 5 × 3
#>   hyperparameters cv_risk stat  
#>   <chr>           <chr>   <chr> 
#> 1 alpha = 0.8     73      min   
#> 2 alpha = 0.8     73      Q1    
#> 3 alpha = 0.6     76      median
#> 4 alpha = 0.3     85      Q3    
#> 5 alpha = 0.1     96      max   
#> 
#>